FastAPI/1.1Introduction to API with Auth/requirements.txt
requests==2.28.2
python-dotenv==1.0.1
argparse==  1.1
fastapi==0.87.0